// superlista-front/src/openapi/model/addRecetaDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AddRecetaDto { 
    menuId: number;
    dia: AddRecetaDto.DiaEnum;
    categoria: AddRecetaDto.CategoriaEnum;
    recetaId: number;
}
export namespace AddRecetaDto {
    export type DiaEnum = 'LUNES' | 'MARTES' | 'MIERCOLES' | 'JUEVES' | 'VIERNES' | 'SABADO' | 'DOMINGO';
    export const DiaEnum = {
        Lunes: 'LUNES' as DiaEnum,
        Martes: 'MARTES' as DiaEnum,
        Miercoles: 'MIERCOLES' as DiaEnum,
        Jueves: 'JUEVES' as DiaEnum,
        Viernes: 'VIERNES' as DiaEnum,
        Sabado: 'SABADO' as DiaEnum,
        Domingo: 'DOMINGO' as DiaEnum
    };
    export type CategoriaEnum = 'DESAYUNO' | 'ALMUERZO' | 'COMIDA' | 'MERIENDA' | 'CENA';
    export const CategoriaEnum = {
        Desayuno: 'DESAYUNO' as CategoriaEnum,
        Almuerzo: 'ALMUERZO' as CategoriaEnum,
        Comida: 'COMIDA' as CategoriaEnum,
        Merienda: 'MERIENDA' as CategoriaEnum,
        Cena: 'CENA' as CategoriaEnum
    };
}